requirements.txt
tqdm
wandb
scipy
timm
scikit-learn
hydra-core
numpy   
debugpy
tensorflow
ipykernel
colorlog
pandas
obspy
umap-learn
omegaconf